refactor(product): Moved persistense directory into infrastructure
