+++ module/product/src/Application/Controller/Api/ProductGridReadAction.php
@@ -14,7 +14,7 @@ use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Grid\Renderer\GridRenderer;
 use Ergonode\Grid\RequestGridConfiguration;
 use Ergonode\Product\Infrastructure\Grid\ProductGrid;
-use Ergonode\Product\Persistence\Dbal\DataSet\DbalProductDataSet;
+use Ergonode\Product\Infrastructure\Persistence\DataSet\DbalProductDataSet;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Swagger\Annotations as SWG;
