+++ module/product/src/Infrastructure/Persistence/DataSet/DbalProductDataSet.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Product\Persistence\Dbal\DataSet;
+namespace Ergonode\Product\Infrastructure\Persistence\DataSet;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\DBAL\Connection;
