+++ module/product/src/Resources/config/services.yml
@@ -4,9 +4,6 @@ services:
         autoconfigure: true
         public: false
 
-    Ergonode\Product\Persistence\:
-        resource: '../../Persistence/*'
-
     Ergonode\Product\Application\:
         resource: '../../Application/*'
         exclude: '../../Domain/{Model}'
@@ -27,20 +24,20 @@ services:
         resource: '../../Infrastructure/Handler/*'
         tags: ['messenger.message_handler']
 
-    Ergonode\Product\Persistence\Dbal\Projector\:
-        resource: '../../Persistence/Dbal/Projector/*'
+    Ergonode\Product\Infrastructure\Persistence\Projector\:
+        resource: '../../Infrastructure/Persistence/Projector/*'
         tags: ['messenger.message_handler']
 
     Ergonode\Product\Infrastructure\JMS\Serializer\Handler\:
         resource: '../../Infrastructure/JMS/Serializer/Handler/*'
         tags: ['jms_serializer.subscribing_handler']
 
-    Ergonode\Product\Domain\Repository\ProductRepositoryInterface: '@Ergonode\Product\Persistence\Dbal\Repository\DbalProductRepository'
-    Ergonode\Product\Domain\Query\ProductQueryInterface: '@Ergonode\Product\Persistence\Dbal\Query\DbalProductQuery'
-    Ergonode\Product\Domain\Query\HistoryQueryInterface: '@Ergonode\Product\Persistence\Dbal\Query\DbalHistoryQuery'
-    Ergonode\Product\Domain\Query\GetProductQueryInterface: '@Ergonode\Product\Persistence\Dbal\Query\GetProductQuery'
+    Ergonode\Product\Domain\Repository\ProductRepositoryInterface: '@Ergonode\Product\Infrastructure\Persistence\Repository\DbalProductRepository'
+    Ergonode\Product\Domain\Query\ProductQueryInterface: '@Ergonode\Product\Infrastructure\Persistence\Query\DbalProductQuery'
+    Ergonode\Product\Domain\Query\HistoryQueryInterface: '@Ergonode\Product\Infrastructure\Persistence\Query\DbalHistoryQuery'
+    Ergonode\Product\Domain\Query\GetProductQueryInterface: '@Ergonode\Product\Infrastructure\Persistence\Query\GetProductQuery'
 
-    Ergonode\Product\Persistence\Dbal\Query\Decorator\CacheProductQueryDecorator:
+    Ergonode\Product\Infrastructure\Persistence\Query\Decorator\CacheProductQueryDecorator:
         decorates: Ergonode\Product\Domain\Query\ProductQueryInterface
         arguments:
-            - '@Ergonode\Product\Persistence\Dbal\Query\Decorator\CacheProductQueryDecorator.inner'
+            - '@Ergonode\Product\Infrastructure\Persistence\Query\Decorator\CacheProductQueryDecorator.inner'
